generator/requirements.txt
python-dotenv>=1.0
numpy>=2.0
pandas>=2.1
scipy>=1.11
sqlalchemy>=2.0
psycopg2-binary>=2.9
boto3>=1.34
pyarrow>=22.0
